+++ src/components/DatePicker/datePickerChangelogData.ts
@@ -10,8 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-07-25",
+    version: "3.2.0",
+    type: "Update",
+    affects: ["Functionality"],
+    notes: ["Exports the `CustomTextInputProps` interface."],
+  },
+  {
+    date: "2024-07-03",
+    version: "3.1.7",
     type: "Update",
     affects: ["Documentation"],
     notes: ["Adds interaction tests for the Controls story."],