+++ src/components/FeaturedContent/FeaturedContent.tsx
@@ -14,17 +14,17 @@ export const featuredContentWidthArray = [
   "twoThirds",
   "threeQuarters",
 ];
-export type featuredContentWidthType = typeof featuredContentWidthArray[number];
+export type FeaturedContentWidthType = typeof featuredContentWidthArray[number];
 
 export const featuredContentPositionArray = ["start", "end"] as const;
-export type featuredContentPositionType =
+export type FeaturedContentPositionType =
   typeof featuredContentPositionArray[number];
 
-interface FeaturedContentImageProps extends ImageProps {
+export interface FeaturedContentImageProps extends ImageProps {
   /** String value that specifies the width of the image rendered within the component. */
-  width?: featuredContentWidthType;
+  width?: FeaturedContentWidthType;
   /** String value that specifies the position of the image rendered within the component. */
-  position?: featuredContentPositionType;
+  position?: FeaturedContentPositionType;
 }
 
 export interface FeaturedContentProps {
@@ -80,11 +80,6 @@ export const FeaturedContent: ChakraComponent<
       if (!imageProps.src) {
         console.warn(
           `NYPL Reservoir FeaturedContent: A value for 'imageProps.src' is required.`
-        );
-      }
-      if (!imageProps.alt) {
-        console.warn(
-          `NYPL Reservoir FeaturedContent: A value for 'imageProps.alt' is required.`
         );
       }
       if (!textContent) {