+++ src/components/MultiSelect/multiSelectChangelogData.ts
@@ -10,8 +10,17 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
-    date: "Prerelease",
-    version: "Prerelease",
+    date: "2024-08-29",
+    version: "3.3.0",
+    type: "New Feature",
+    affects: ["Accessibility"],
+    notes: [
+      "Renders the `buttonText` prop as the `CheckboxGroup`'s legend for more meaningful context.",
+    ],
+  },
+  {
+    date: "2024-07-03",
+    version: "3.1.7",
     type: "New Feature",
     affects: ["Functionality"],
     notes: [